| update | nothing to be done |

+++ vm_m/libft/srcs/ft_is_integer.c
@@ -6,7 +6,7 @@
 /*   By: yhetman <yhetman@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/10/26 14:07:40 by yhetman           #+#    #+#             */
-/*   Updated: 2019/10/26 14:12:23 by yhetman          ###   ########.fr       */
+/*   Updated: 2019/10/27 14:18:55 by blukasho         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -24,15 +24,15 @@ static size_t	ft_strlen_of_word(const char *str)
 
 int				ft_is_integer(const char *arg)
 {
-	intmax_t	tmp;
+	long		tmp;
 	size_t		len;
 
 	len = ft_strlen_of_word(arg);
 	if (len > 11)
 		return (0);
 	tmp = ft_atoi(arg);
-	if ((intmax_t)INT_MAX < (intmax_t)tmp
-		|| (intmax_t)tmp < (intmax_t)INT_MIN)
+	if ((long)INT_MAX < tmp
+		|| tmp < (long)INT_MIN)
 		return (0);
 	if (*arg == '+' || *arg == '-')
 		arg++;
